requirements.txt
python-telegram-bot==20.3
qrcode[pil]==7.4.2
Pillow==10.3.0
reportlab==4.1.0
